backend/requirements.txt
# fastapi 
# uvicorn 
# langchain 
# langgraph 
# langchain-groq 
# python-dotenv 
# pydantic
# langchain-community
# langchain-huggingface 
# arxiv
# unstructured
# langchain-community 
# langchain-huggingface 
# faiss-cpu
# sentence-transformers
# youtube-transcript-api
# requests
# beautifulsoup4
# react-markdown
# remark-gfm

aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
arxiv==2.4.0
attrs==25.4.0
backoff==2.2.1
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.4
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
emoji==2.15.0
faiss-cpu==1.13.2
fastapi==0.128.5
feedparser==6.0.12
filelock==3.20.3
filetype==1.2.0
frozenlist==1.8.0
fsspec==2026.2.0
graphviz==0.21
greenlet==3.3.1
groq==0.37.1
h11==0.16.0
hf-xet==1.2.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.4.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.9
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.9
langchain-groq==1.1.2
langchain-huggingface==1.2.0
langchain-text-splitters==1.1.0
langdetect==1.0.9
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.4
langsmith==0.6.9
llvmlite==0.46.0
lxml==6.0.2
MarkupSafe==3.0.3
marshmallow==3.26.2
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
networkx==3.6.1
nltk==3.9.2
numba==0.63.1
numpy==2.3.5
olefile==0.47
orjson==3.11.7
ormsgpack==1.12.2
packaging @ file:///C:/miniconda3/conda-bld/packaging_1761049096285/work
propcache==0.4.1
psutil==7.2.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyMySQL==1.1.2
pypdf==6.7.0
pypdfium2==5.4.0
python-dotenv==1.2.1
python-iso639==2026.1.31
python-magic==0.4.27
python-multipart==0.0.22
python-oxmsg==0.0.2
PyYAML==6.0.3
RapidFuzz==3.14.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.2
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.46
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==5.1.0
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
unstructured==0.18.31
unstructured-client==0.42.10
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
webencodings==0.5.1
wrapt==2.1.1
xxhash==3.6.0
yarl==1.22.0
youtube-transcript-api==1.2.4
zstandard==0.25.0
